requirements.txt
opencv-python
face_recognition
git+https://github.com/ageitgey/face_recognition_models
scikit-learn
Pillow
numpy
